Nit fixes.
