+++ src/buffer/query.rs
@@ -1,8 +1,9 @@
 use std::path::PathBuf;
 
 use data::dashboard::BufferAction;
+use data::preview::{self, Previews};
 use data::target::{self, Target};
-use data::{Config, Server, buffer, history, message, preview};
+use data::{Config, Server, buffer, history, message};
 use iced::advanced::text;
 use iced::widget::{column, container, row, vertical_space};
 use iced::{Length, Task};
@@ -50,7 +51,7 @@ pub fn view<'a>(
     let chathistory_state =
         clients.get_chathistory_state(server, &query.to_target());
 
-    let previews = Some(buffer::Previews::from(
+    let previews = Some(Previews::new(
         previews,
         &query.to_target(),
         &config.preview,
